feat(cms-sections): drop JSON tab, keep preview-only side pane
Removes the JSON editor feature from the section side panel — it wasn't a
satisfying experience. The right-hand pane is now preview-only: collapsible,
with device (desktop/mobile) toggle, refresh, and open-in-new-tab. Saved
sections open it by default; available (unsaved) sections keep it collapsed
and auto-reload the preview (debounced) as the form changes.
